new option 'HELPDESK_SHOW_EDIT_BUTTON_FOLLOW_UP' which allows superusers to delete individual followups (so you don't have to go to the admin). (by @kotowicz in a2a5b10b2d44c861d2b0732660fb2736ee65e31b)

+++ helpdesk/views/staff.py
@@ -143,7 +143,7 @@ def delete_ticket(request, ticket_id):
         return HttpResponseRedirect(reverse('helpdesk_home'))
 delete_ticket = staff_member_required(delete_ticket)
 
-def followup_edit(request, ticket_id, followup_id, ):
+def followup_edit(request, ticket_id, followup_id):
     "Edit followup options with an ability to change the ticket."
     followup = get_object_or_404(FollowUp, id=followup_id)
     ticket = get_object_or_404(Ticket, id=ticket_id)
@@ -155,7 +155,7 @@ def followup_edit(request, ticket_id, followup_id, ):
                                       'public': followup.public,
                                       'new_status': followup.new_status,
                                       })
-        
+    
         return render_to_response('helpdesk/followup_edit.html',
             RequestContext(request, {
                 'followup': followup,
@@ -178,14 +178,27 @@ def followup_edit(request, ticket_id, followup_id, ):
                 new_followup.user = followup.user
             new_followup.save()
             # get list of old attachments & link them to new_followup
-            attachments = Attachment.objects.filter(followup = followup)            
+            attachments = Attachment.objects.filter(followup = followup)
             for attachment in attachments:
                 attachment.followup = new_followup
                 attachment.save()
             # delete old followup
-            followup.delete()                
+            followup.delete()
         return HttpResponseRedirect(reverse('helpdesk_view', args=[ticket.id]))
-            
+
+
+def followup_delete(request, ticket_id, followup_id):
+    ''' followup delete for superuser'''
+
+    ticket = get_object_or_404(Ticket, id=ticket_id)
+    if not request.user.is_superuser:
+        return HttpResponseRedirect(reverse('helpdesk_view', args=[ticket.id]))
+
+    followup = get_object_or_404(FollowUp, id=followup_id)
+    followup.delete()
+    return HttpResponseRedirect(reverse('helpdesk_view', args=[ticket.id]))
+
+
 def view_ticket(request, ticket_id):
     ticket = get_object_or_404(Ticket, id=ticket_id)
